% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{account_login}
\alias{account_login}
\title{Log into your LOBSTER account}
\usage{
account_login(login, pwd)
}
\arguments{
\item{login}{Your unique account user email address}

\item{pwd}{Your unique account password}
}
\value{
An account object which contains the relevant session data.
}
\description{
Log into your LOBSTER account
}
