% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{compute_depth}
\alias{compute_depth}
\title{compute_depth}
\usage{
compute_depth(orderbook, side = "bid", bp = 0)
}
\arguments{
\item{orderbook}{The orderbook file}

\item{side}{The side for which depth should be computed}

\item{bp}{Basis points from best quote}
}
\value{
blablabla
}
\description{
compute_depth
}
