% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{.account_archive}
\alias{.account_archive}
\title{Fetch LOBSTER archive data}
\usage{
.account_archive(account_login)
}
\arguments{
\item{account_login}{output of the [.account_login] function}
}
\value{
Tibble with archive data.
\itemize{
  \item symbol [character]
  \item start_date [Date]
  \item end_date [Date]
  \item level [integer]
  \item size [integer]
  \item download [character]
  \item id [integer]
}
}
\description{
Fetch LOBSTER archive data
}
