% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{.process_collect}
\alias{.process_collect}
\title{.process_collect}
\usage{
.process_collect(path, clean_up = TRUE)
}
\arguments{
\item{path}{The path of the files}

\item{clean_up}{Remove the raw files after collecting lobsterdata}
}
\value{
blablabla
}
\description{
.process_collect
}
