% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_data}
\alias{process_data}
\title{Processes lobster files and returns a clean, zipped csv file}
\usage{
process_data(path, clean_up = TRUE)
}
\arguments{
\item{path}{The path of the files}

\item{clean_up}{TRUE Remove the raw files after collecting lobsterdata}
}
\description{
Processes lobster files and returns a clean, zipped csv file
}
