% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{account_archive}
\alias{account_archive}
\title{Fetch LOBSTER archive data}
\usage{
account_archive(account_login)
}
\arguments{
\item{account_login}{output of the \link{account_login} function}
}
\value{
Tibble with archive data.
\itemize{
\item symbol \link{character}
\item start_date \link{Date}
\item end_date \link{Date}
\item level \link{integer}
\item size \link{integer}
\item download \link{character}
\item id \link{integer}
}
}
\description{
Fetch LOBSTER archive data
}
