% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{account_login}
\alias{account_login}
\title{Log into your LOBSTER account}
\usage{
account_login(login, pwd)
}
\arguments{
\item{login}{user email address}

\item{pwd}{user password}
}
\value{
List with session data.
}
\description{
Log into your LOBSTER account
}
