% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{data_download}
\alias{data_download}
\title{Download requested data}
\usage{
data_download(requested_data, account_login, path = ".", unzip = TRUE)
}
\arguments{
\item{requested_data}{Tibble with archive data.}

\item{account_login}{An account object which contains the relevant session data.}

\item{path}{relative path where the data should be stored}

\item{unzip}{TRUE If true, the .7z files are automatically unzipped. This can be omitted.}
}
\description{
Download requested data
}
