% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{.request_validate}
\alias{.request_validate}
\title{Validate a request}
\usage{
.request_validate(request_query, account_archive = NULL)
}
\arguments{
\item{request_query}{A tibble which states symbol, start_date, end_date, and level of the required data}

\item{account_archive}{NULL If provided, the validation filters out data which is already available in the account archive}
}
\description{
Validate a request
}
