% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request_submit}
\alias{request_submit}
\title{Submit a request}
\usage{
request_submit(account_login, request)
}
\arguments{
\item{account_login}{An account object which contains the relevant session data.}

\item{request}{A tibble with the requested trading days.}
}
\value{
A tibble with an overview of the requested trading days.
}
\description{
Submit a request
}
